pyproject.toml
[tool.poetry]
name = "figflix"
version = "1.0.0"
description = "A Django-based movie web application with TMDb API integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14"
django-cors-headers = "^4.3"
python-decouple = "^3.8"
requests = "^2.31"
pillow = "^10.1"
djangorestframework-simplejwt = "^5.3"

[tool.poetry.group.dev.dependencies]
black = "^23.11"
flake8 = "^6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
